router.py
import importlib
import requests
import schedule
import time
import json
from typing import List, Dict, Any, Tuple
from datetime import datetime
import threading
from huggingface_hub import HfApi
from github import Github
import numpy as np

class DubIndex:
    """
    Implementation of the DubIndex scoring system for LLM evaluation.
    DubIndex = 0.20×[Speed] + 0.30×[Hallucination Resistance] + 0.30×[Accuracy] + 
               0.10×[Efficiency] + 0.10×[Size Efficiency]
    """
    @staticmethod
    def calculate(
        speed: float,
        hallucination_resistance: float,
        accuracy: float,
        efficiency: float,
        model_size: float,
        speed_range: Tuple[float, float],
        accuracy_range: Tuple[float, float],
        efficiency_range: Tuple[float, float]
    ) -> float:
        # Normalize metrics
        speed_norm = (speed - speed_range[0]) / (speed_range[1] - speed_range[0])
        hallucination_norm = 1 / (1 + hallucination_resistance)  # Lower is better
        accuracy_norm = (accuracy - accuracy_range[0]) / (accuracy_range[1] - accuracy_range[0])
        efficiency_norm = (efficiency - efficiency_range[0]) / (efficiency_range[1] - efficiency_range[0])
        size_efficiency = 1 / (1 + model_size)  # Lower is better

        # Calculate DubIndex
        dub_index = (
            0.20 * speed_norm +
            0.30 * hallucination_norm +
            0.30 * accuracy_norm +
            0.10 * efficiency_norm +
            0.10 * size_efficiency
        )
        
        return round(dub_index, 2)

class ModelDiscovery:
    """Handles automatic discovery of new LLM models."""
    def __init__(self):
        self.hf_api = HfApi()
        self.github_api = Github()  # Use token in production
        self.known_models = set()

    async def scan_huggingface(self) -> List[Dict[str, Any]]:
        """Scan HuggingFace for new LLM models."""
        models = []
        for model in self.hf_api.list_models(filter="text-generation"):
            if model.id not in self.known_models:
                models.append({
                    "name": model.id,
                    "source": "huggingface",
                    "url": f"https://huggingface.co/{model.id}",
                    "downloads": model.downloads,
                    "likes": model.likes
                })
        return models

    async def scan_github(self) -> List[Dict[str, Any]]:
        """Scan GitHub for new LLM implementations."""
        models = []
        query = "language:Python topic:llm"
        repos = self.github_api.search_repositories(query, sort="stars")
        
        for repo in repos:
            if repo.full_name not in self.known_models:
                models.append({
                    "name": repo.full_name,
                    "source": "github",
                    "url": repo.html_url,
                    "stars": repo.stargazers_count,
                    "last_updated": repo.updated_at
                })
        return models

class ModelEvaluator:
    """Evaluates LLM models using the DubIndex scoring system."""
    def __init__(self):
        self.benchmark_tasks = [
            "Explain quantum computing",
            "Debug this Python code",
            "Write a creative story",
            "Analyze this research paper"
        ]
        
    async def evaluate_model(self, model: Dict[str, Any]) -> float:
        """
        Evaluate a model's performance using the DubIndex metrics.
        Returns the DubIndex score.
        """
        # Simulate model evaluation - replace with actual benchmarking in production
        speed = np.random.uniform(100, 1000)  # tokens/sec
        hallucination_rate = np.random.uniform(0.01, 0.1)
        accuracy = np.random.uniform(0.8, 0.99)
        efficiency = np.random.uniform(0.7, 0.95)
        model_size = np.random.uniform(1, 100)  # billions of parameters

        return DubIndex.calculate(
            speed=speed,
            hallucination_resistance=hallucination_rate,
            accuracy=accuracy,
            efficiency=efficiency,
            model_size=model_size,
            speed_range=(100, 1000),
            accuracy_range=(0.8, 0.99),
            efficiency_range=(0.7, 0.95)
        )

class AdaptiveLLMRouter:
    """
    Advanced LLM router with automatic model discovery, evaluation,
    and dynamic switching based on performance.
    """
    def __init__(self):
        self.discovery = ModelDiscovery()
        self.evaluator = ModelEvaluator()
        self.models: Dict[str, Dict[str, Any]] = {}
        self.task_mappings: Dict[str, str] = {}
        
        # Start periodic scanning
        self.start_scanning()

    def start_scanning(self):
        """Start periodic scanning for new models."""
        schedule.every(72).hours.do(self.scan_and_evaluate)
        
        # Run scanning in a separate thread
        def run_scheduler():
            while True:
                schedule.run_pending()
                time.sleep(1)
                
        threading.Thread(target=run_scheduler, daemon=True).start()

    async def scan_and_evaluate(self):
        """Scan for new models and evaluate them."""
        print(f"Starting model scan at {datetime.now()}")
        
        # Discover new models
        new_models = []
        new_models.extend(await self.discovery.scan_huggingface())
        new_models.extend(await self.discovery.scan_github())
        
        # Evaluate new models
        for model in new_models:
            score = await self.evaluator.evaluate_model(model)
            model["dub_index"] = score
            
            # If model performs better than existing ones, add it to rotation
            if self.should_add_model(model):
                self.models[model["name"]] = model
                self.update_task_mappings(model)
                
        self.save_state()

    def should_add_model(self, model: Dict[str, Any]) -> bool:
        """Determine if a new model should be added based on its performance."""
        if not self.models:
            return True
            
        avg_score = np.mean([m["dub_index"] for m in self.models.values()])
        return model["dub_index"] > avg_score

    def update_task_mappings(self, model: Dict[str, Any]):
        """Update task mappings based on model performance."""
        # Example mapping logic - extend based on actual model capabilities
        if model["dub_index"] > 0.9:
            self.task_mappings["general"] = model["name"]
        if model["dub_index"] > 0.85:
            self.task_mappings["coding"] = model["name"]
        if model["dub_index"] > 0.8:
            self.task_mappings["creative"] = model["name"]

    def save_state(self):
        """Save current state to disk."""
        state = {
            "models": self.models,
            "task_mappings": self.task_mappings,
            "last_updated": datetime.now().isoformat()
        }
        with open("llm_router_state.json", "w") as f:
            json.dump(state, f)

    async def route_task(self, task: str, context: Dict[str, Any]) -> str:
        """Route a task to the most appropriate model."""
        task_lower = task.lower()
        
        # Find the best model for the task
        for task_type, model_name in self.task_mappings.items():
            if task_type in task_lower:
                model = self.models.get(model_name)
                if model:
                    return f"Using {model_name} (DubIndex: {model['dub_index']}) for task"
        
        # Use general-purpose model as fallback
        general_model = self.task_mappings.get("general")
        if general_model:
            return f"Using general model {general_model} for task"
            
        return "No suitable model found for task"
