requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
redis>=5.0.1
pydantic>=2.4.2
requests>=2.31.0
numpy>=1.24.0
schedule>=1.2.0

# API integrations
huggingface-hub>=0.19.0
PyGithub>=2.1.1

# Development dependencies
build>=1.0.3
twine>=4.0.2
setuptools>=68.0.0
wheel>=0.41.2

# Type checking
types-setuptools>=68.2.0.0
types-redis>=4.6.0.20231023
types-requests>=2.31.0.10
